Fixed code smells
