--- src/main/java/com/revature/controller/FlightController.java
@@ -1,13 +1,0 @@
-package com.revature.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.server.WebSession;
-
-import com.revature.beans.Flight;
-
-import reactor.core.publisher.Flux;
-
-public interface FlightController {
-	
-	public ResponseEntity<Flux<Flight>> getFlightsByDestination(String destination, WebSession session);
-}
