--- src/main/java/com/revature/controller/HotelController.java
@@ -1,13 +1,0 @@
-package com.revature.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.server.WebSession;
-
-import com.revature.beans.Hotel;
-
-import reactor.core.publisher.Flux;
-
-public interface HotelController {
-	
-	public ResponseEntity<Flux<Hotel>> getHotelsByLocation(String location, WebSession session);
-}
