+++ src/main/java/com/revature/controllers/ReservationController.java
@@ -3,9 +3,6 @@ package com.revature.controllers;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.server.WebSession;
 
-import com.revature.beans.Car;
-import com.revature.beans.Flight;
-import com.revature.beans.Hotel;
 import com.revature.beans.Reservation;
 
 import reactor.core.publisher.Mono;
