+++ src/main/java/com/revature/controllers/UserControllerImpl.java
@@ -22,6 +22,7 @@ import com.revature.beans.User;
 import com.revature.beans.Vacation;
 import com.revature.services.UserService;
 
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 @RestController
@@ -65,7 +66,7 @@ public class UserControllerImpl implements UserController {
 	public Mono<ResponseEntity<User>> register(@RequestBody User u, @PathVariable("username") String username) {
 		// check to see if that username is available
 		return userService.checkAvailability(username).flatMap(b -> {
-			if (!b) {
+			if (Boolean.FALSE.equals(b)) {
 				return userService.register(username, u.getPassword(), u.getEmail(), u.getFirstName(), u.getLastName(),
 						u.getBirthday(), u.getType()).map(user -> ResponseEntity.status(201).body(user));
 			} else {
@@ -115,6 +116,18 @@ public class UserControllerImpl implements UserController {
 				return ResponseEntity.ok(v);
 			}
 		});
+	}
+	
+	@VacationerCheck
+	@DeleteMapping("{username}")
+	public Mono<ResponseEntity<Void>> deleteUser(@PathVariable("username") String username, WebSession session) {
+		User loggedUser = (User) session.getAttribute(LOGGED_USER);
+		loggedUser = loggedUser == null ? new User() : loggedUser;
+		session.invalidate();
+		return Flux.from(userService.login(username, loggedUser.getPassword())).map(u -> u.getVacations())
+				.flatMap(l -> Flux.fromIterable(l)).flatMap(uuid -> userService.getVacation(username, uuid))
+				.collectList().map(list -> userService.deleteUser(username, list))
+				.map(v -> ResponseEntity.status(204).build());
 	}
 
 }
