+++ src/main/java/com/revature/data/UserDao.java
@@ -3,7 +3,6 @@ package com.revature.data;
 import org.springframework.data.cassandra.repository.ReactiveCassandraRepository;
 import org.springframework.stereotype.Repository;
 
-import com.revature.beans.User;
 import com.revature.dto.UserDto;
 
 import reactor.core.publisher.Mono;
