+++ src/test/java/com/revature/services/CarServiceTest.java
@@ -13,10 +13,8 @@ import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
 import com.revature.beans.Car;
-import com.revature.beans.Hotel;
 import com.revature.data.CarDao;
 import com.revature.dto.CarDto;
-import com.revature.dto.HotelDto;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
