+++ src/test/java/com/revature/services/UserServiceTests.java
@@ -2,7 +2,6 @@ package com.revature.services;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.mockito.Mockito.when;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -374,7 +373,7 @@ class UserServiceTests {
 	}
 	
 	@Test
-	void testdeleteUser() {
+	void testDeleteUser() {
 		
 		 List<Vacation> vacList = new ArrayList<Vacation>();
 		 vacList.add(vac);
